# openapi.json
{
  "components": {
    "schemas": {
...
      "WorkflowPlanCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "expires_in_minutes": {
            "default": 30,
            "maximum": 60,
            "minimum": 1,
            "title": "Expires In Minutes",
            "type": "integer"
          },
          "overrides": {
            "additionalProperties": true,
            "title": "Overrides",
            "type": "object"
          },
          "target_stage": {
            "default": "generate_audio",
            "pattern": "^[a-z][a-z0-9_]{0,79}$",
            "title": "Target Stage",
            "type": "string"
          }
        },
        "title": "WorkflowPlanCreateRequest",
        "type": "object"
      },
      "WorkflowPlanExecuteRequest": {
        "additionalProperties": false,
        "properties": {
          "accepted_confirmations": {
            "items": {
              "type": "string"
            },
            "title": "Accepted Confirmations",
            "type": "array"
          },
          "plan_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Plan Digest",
            "type": "string"
          }
        },
        "required": [
          "plan_digest"
        ],
        "title": "WorkflowPlanExecuteRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerToken": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      },
      "nativeOAuth": {
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/v1/auth/automation/authorize",
            "scopes": {
              "app.cancel": "Cancel durable application work.",
              "app.credentials.read": "Inspect credential status.",
              "app.credentials.write": "Change credential references.",
              "app.read": "Read application state.",
              "app.run": "Start durable application work.",
              "app.write": "Change reversible application state.",
              "manager.mutate": "Execute approved Manager plans.",
              "manager.read": "Inspect Manager state and plans.",
              "manager.runtime": "Control managed runtime services."
            },
            "tokenUrl": "/api/v1/auth/automation/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Pandrator API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions/{sessionId}/workflow-plans": {
      "post": {
        "operationId": "createWorkflowPlan",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPlanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Immutable workflow execution preview"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerToken": []
          },
          {
            "nativeOAuth": [
              "app.read"
            ]
          }
        ]
      }
    },
...
    "/api/v1/workflow-plans/{planId}": {
      "get": {
        "operationId": "getWorkflowPlan",
        "parameters": [
          {
            "in": "path",
            "name": "planId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable workflow execution preview"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerToken": []
          },
          {
            "nativeOAuth": [
              "app.read"
            ]
          }
        ]
      }
    },
    "/api/v1/workflow-plans/{planId}/execute": {
      "post": {
        "operationId": "executeWorkflowPlan",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "planId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPlanExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Exact plan consumed and queued"
          },
          "409": {
            "description": "Plan stale, consumed, expired, or confirmation missing"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerToken": []
          },
          {
            "nativeOAuth": [
              "app.run"
            ]
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}